重新生成档案服务实体模型

+++ Services/sql/archive/t.branch_group.sql
@@ -1,7 +1,7 @@
 DROP TABLE IF EXISTS `branch_group`;
 
 CREATE TABLE `branch_group` (
-    `id` BIGINT NOT NULL AUTO_INCREMENT COMMENT '自增编码',
+    `id` INT NOT NULL AUTO_INCREMENT COMMENT '自增编码',
     `name` VARCHAR(45) NOT NULL COMMENT '名称',
     PRIMARY KEY (`id`))
 ENGINE = InnoDB
